Attempt to fix exit in SSM agent module
Failed to stop amazon-ssm-agent.service: Unit amazon-ssm-agent.service not loaded.

I think this may still fail, because this will return 4, and that may cause packer to
exist. Last resort is to disable exit on error temporarily.

+++ scripts/install_ssm_agent.sh
@@ -2,9 +2,16 @@
 set -o errexit -o nounset -o pipefail
 
 echo ">>> Installing latest SSM agent"
-systemctl stop amazon-ssm-agent
-rm -rf /var/log/amazon/ssm/*
-yum erase -y amazon-ssm-agent
+systemctl status amazon-ssm-agent
+if [ $? = 0 ]
+then
+    echo ">>>> SSM agent installed. Stop and remove it"
+    systemctl stop amazon-ssm-agent
+    rm -rf /var/log/amazon/ssm/*
+    yum erase -y amazon-ssm-agent
+else
+    echo ">>>> SSM agent not yet installed"
+fi
 yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
 systemctl enable amazon-ssm-agent
